+++ Core/Util/Constants.cs
@@ -303,4 +303,6 @@ public static class Constants
     public const int MaxTextureHeight = 16384;
 
     public static readonly int MaxSoundChannels = Enum.GetValues<SoundChannel>().Length;
+
+    public const int DefaultMaxDistance = 6000;
 }